moved part of settings to dedicated file
